% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiForecast.R
\name{MultiForecast}
\alias{MultiForecast}
\title{Generate multiple Time Series forecasts}
\usage{
MultiForecast(ts, rdm = c(0, 1), act, fper, type = "additive",
  mod = c("aa", "sm", "hw"), weight = F, xreg = NULL, freg = NULL)
}
\arguments{
\item{ts}{The Time Series object to be used for forecasting}

\item{rdm}{The distinct numerical values to be used in combination as p,d,q (trend component) & P,D,Q (seasonal component) for the ARIMA models}

\item{act}{The actual values for the validation period based on which the models will be evaluated}

\item{fper}{The forecasting period}

\item{type}{The type of Time Series being analysed ("additive" or "multiplicative")}

\item{mod}{The different models to be used for forecasting ("aa" for auto.arima,"sm" for Simulated ARIMA,"hw" for HoltWinters)}

\item{weight}{If the ensembled forecasts should be generated based on a weighted average as per the errors from each model instead of a simple mean}

\item{xreg}{Data for external regressors to be used for forecasting}

\item{freg}{Data for external regressors for the validation period}
}
\value{
A list containing the forecasted values from each model along with the ensembled forecast
}
\description{
Uses a time series object as input and generates multiple forecasts using different models to provide an ensembled forecast
}
